# e2e/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Conduit API",
      "version": "1.0.0",
      "title": "Conduit API",
      "contact": {
        "name": "RealWorld",
        "url": "https://realworld.io"
      },
      "license": {
        "name": "MIT License",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "basePath": "/api",
    "schemes": [
      "https",
      "http"
    ],
    "produces": [
      "application/json"
    ],
    "consumes": [
      "application/json"
    ],
    "securityDefinitions": {
      "Token": {
        "description": "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token: xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "paths": {
...
      "/profiles/{username}": {
        "get": {
          "summary": "Get a profile",
          "description": "Get a profile of a user of the system. Auth is optional",
          "tags": [
            "Profile"
          ],
          "operationId": "GetProfileByUsername",
          "parameters": [
            {
              "name": "username",
              "in": "path",
              "description": "Username of the profile to get",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/ProfileResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/profiles/{username}/follow": {
        "post": {
          "summary": "Follow a user",
          "description": "Follow a user by username",
          "tags": [
            "Profile"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "FollowUserByUsername",
          "parameters": [
            {
              "name": "username",
              "in": "path",
              "description": "Username of the profile you want to follow",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/ProfileResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "delete": {
          "summary": "Unfollow a user",
          "description": "Unfollow a user by username",
          "tags": [
            "Profile"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "UnfollowUserByUsername",
          "parameters": [
            {
              "name": "username",
              "in": "path",
              "description": "Username of the profile you want to unfollow",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/ProfileResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "profile": {
            "$ref": "#/definitions/Profile"
          }
        },
        "required": [
          "profile"
        ]
      },
      "Profile": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "following": {
            "type": "boolean"
          }
        },
        "required": [
          "username",
          "bio",
          "image",
          "following"
        ]
      },
...
      "GenericErrorModel": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "body": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "body"
            ]
          }
        },
        "required": [
          "errors"
        ]
      }
    }
  }